Add files via upload
This java file makes the game "Rescue the Queens"
The buttons all appear as "*" at first and when clicked on
they are shown to be queens or blanks. If they are blanks
the all blanks around them are reveal and so on in a cascading effect.
You win if all queens are revealed before all of the blanks are.